Move updating answers to QuestionDef
